requirements.txt
PyOpenGL
pylint